dhcpd.conf: force DHCP-specified hostname to DDNS for Galil/Camera
For Galil the client-provided hostname
is not strictly valid as it contains a
nbsp character (\255).

For Camera there is just no provided
hostname whatsoever.

For both of these cases it's better
to just use our naming convention
and set the names at the DHCP scope
and just foward the names to DNS when
a DHCP request arrives.